Correct errors to enable baremetal simulation
Various errors were hidden in the makefile or in the source code.
They were corrected and now we can simulate without UART and no
portable initialization (it is saved for later).

This commit is meant to be used to check the modified files in
order to reproduce a good working environment.